poridhi/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
numpy
pandas
datasets
jinja2
redis
locust